tiles/requirements.txt
#torch
#torchvision
matplotlib
Pillow
click
Faker